requirements.txt
numpy<1.19.0
pandas
PyYAML
matplotlib
tensorflow
tensorflow_datasets
sklearn
transformers
seaborn
torch
